WIP: reconstruct KCLayout to be Pdk and old KCLayout
